Update temporary code to own config
