+++ src/backend/efiling-reviewer-api/src/test/java/ca/bc/gov/open/jag/efilingreviewerapi/cache/CacheConfigTest.java
@@ -1,11 +1,18 @@
 package ca.bc.gov.open.jag.efilingreviewerapi.cache;
 
+import ca.bc.gov.open.efilingdiligenclient.diligen.*;
+import ca.bc.gov.open.jag.efilingdiligenclient.api.AuthenticationApi;
+import ca.bc.gov.open.jag.efilingdiligenclient.api.DocumentsApi;
+import ca.bc.gov.open.jag.efilingreviewerapi.extract.mappers.ExtractRequestMapper;
+import ca.bc.gov.open.jag.efilingreviewerapi.extract.store.ExtractStore;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.data.redis.cache.RedisCacheManager;
+import org.springframework.web.client.RestTemplate;
 
 import static org.assertj.core.api.Assertions.assertThat;
